Arrays Commit
- Added the Arrays.java class, which showcases the various ways of
declaring and using different types of Arrays.
- Updated comments on a few files.
- In the previous commit, I just added a WaitForKey.java class that
allows the calling of a function which use is to wait for the user to
pres the Enter key. I just forgot to mention it on the previous
description (my bad, whoops!).
- Updated the method names in every class so they read "Class.showcase".
This is logical because it will allow to use these classes for another
purpose in the future (if We ever need to use them!).
- Added the proper Array calling line at the Controller class.